+++ apps/api/src/mf/chat/stream/rag-stream.ts
@@ -2,84 +2,13 @@ import { Response } from 'express'
 import { Response as FetchResponse } from 'node-fetch'
 import { ChatRecordStatus, UpdateTarget } from '../types/interfaces.js'
 import { logger } from '../../../logger.js'
-import { CONFIG } from '../config/constants.js'
 import { formatErrorMessage } from '../../../utils/format.js'
 import { prisma } from '@briefer/database'
 import { activeRequests } from '../utils/fetch.js'
 import path from 'path'
 import fs from 'fs/promises'
-import { APIError, ERROR_CODES } from '../types/errors.js' // Import APIError
-
-// 设置SSE连接
-export function setupSSEConnection(res: Response) {
-  res.setHeader('Content-Type', 'text/event-stream')
-  res.setHeader('Cache-Control', 'no-cache')
-  res.setHeader('Connection', 'keep-alive')
-  res.flushHeaders()
-
-  // 设置保活定时器
-  const keepAliveTimer = setInterval(() => {
-    res.write(': keepalive\n\n')
-  }, CONFIG.SSE.KEEP_ALIVE_INTERVAL)
-
-  // 清理函数
-  const cleanup = () => {
-    clearInterval(keepAliveTimer)
-    res.end()
-  }
-
-  // 监听连接关闭
-  res.on('close', cleanup)
-}
-
-// 发送SSE错误
-export async function sendSSEError(res: Response, error: unknown, updateTarget?: UpdateTarget) {
-  const formattedError = formatErrorMessage(error)
-
-  // 如果存在更新目标，将错误消息保存到数据库
-  if (updateTarget?.type === 'chat_record' && updateTarget.id) {
-    try {
-      await prisma().$transaction([
-        prisma().chatRecord.update({
-          where: { id: updateTarget.id },
-          data: {
-            answer: Buffer.from(formattedError),
-            speakerType: 'assistant',
-            status: ChatRecordStatus.ERROR,
-            updateTime: new Date(),
-          },
-        }),
-        prisma().chat.update({
-          where: { id: updateTarget.chatId },
-          data: { updateTime: new Date() },
-        }),
-      ])
-    } catch (dbError) {
-      logger().error({
-        msg: 'Failed to save error message to database',
-        data: { error: dbError },
-      })
-    }
-  }
-
-  // 分行发送错误消息，确保格式正确
-  formattedError.split('\n').forEach((line) => {
-    res.write(`data: ${line}\n`)
-  })
-  res.write('\n') // 表示该消息结束
-  res.write('data: [DONE]\n\n')
-  // res.end() // 统一不关闭
-}
-
-// 发送SSE消息
-export function sendSSEMessage(res: Response, data: any, type: string = 'message') {
-  const message = {
-    type,
-    data
-  }
-
-  res.write(`data: ${JSON.stringify(message)}\n\n`)
-}
+import { APIError, ERROR_CODES } from '../types/errors.js'
+import { sendSSEError } from '../utils/sse-utils.js'
 
 // 处理流式响应
 export async function handleStreamResponse(
@@ -423,70 +352,4 @@ export async function handleStreamResponse(
       activeRequests.delete(updateTarget.roundId)
     }
   }
-}
-
-// 创建SSE响应
-export function createSSEResponse(data: any): string {
-  return `data: ${JSON.stringify(data)}\n\n`
-}
-
-// 检查SSE支持
-export function checkSSESupport(req: Request): boolean {
-  return req.headers.get('accept') === 'text/event-stream'
-}
-
-// 发送SSE完成消息
-export function sendSSEComplete(res: Response, updateTarget?: UpdateTarget) {
-  const completeData = {
-    type: 'complete',
-    data: {
-      message: 'Stream completed',
-      ...(updateTarget && { updateTarget })
-    }
-  }
-
-  res.write(`data: ${JSON.stringify(completeData)}\n\n`)
-  res.end()
-}
-
-// SSE相关工具函数
-export type SSEUpdateTarget = {
-  type: 'chat_record' | 'chat'
-  chatId: string
-  roundId?: string
-}
-
-export const setupSSEConnectionUtil = (res: Response) => {
-  res.setHeader('Content-Type', 'text/event-stream')
-  res.setHeader('Cache-Control', 'no-cache')
-  res.setHeader('Connection', 'keep-alive')
-  res.flushHeaders()
-}
-
-export const sendSSEMessageUtil = (res: Response, data: any) => {
-  res.write(`data: ${JSON.stringify(data)}\n\n`)
-}
-
-export const sendSSEErrorUtil = async (res: Response, error: Error, target: SSEUpdateTarget) => {
-  logger().error({
-    msg: 'SSE error',
-    data: {
-      error: error.message,
-      target,
-    },
-  })
-
-  if (target.type === 'chat_record' && target.roundId) {
-    await prisma().chatRecord.update({
-      where: {
-        id: target.roundId,
-      },
-      data: {
-        status: ChatRecordStatus.ERROR,
-        updateTime: new Date(),
-      },
-    })
-  }
-
-  res.end()
 }
